requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-autoslug>=1.9.9
django-taggit>=5.0.1
python-decouple>=3.8
Pillow>=10.2.0
django-markdownx
drf-spectacular
Djoser
social-auth-app-Django
gunicorn
dj_database_url                                                                                              